+++ doc/changelog.rst
@@ -41,6 +41,7 @@ Xtheme
 
 - Add option to render extra templates in theme configuration.
 - Add option for per object placeholders
+- Fix bug deleting last row from placeholder
 
 Shuup 1.6.6
 -----------